+++ .gitignore
@@ -11,7 +11,6 @@ dist/
 downloads/
 eggs/
 .eggs/
-lib/
 lib64/
 parts/
 sdist/
@@ -102,8 +101,6 @@ dmypy.json
 .app
 
 # Configurations
-config.yaml
-config.yml
 
 #Documents
 .all-docs
