+++ docker-compose.yml
@@ -1,22 +1,41 @@
 version: '3.7'
 
 services:
-  qdrant:
-    image: qdrant/qdrant
+  my-api:
+    build: 
+      context: .
     ports:
-      - "6333:6333" 
-    networks:
-      - my_network
-
-  fastapi:
-    build: .  # Path to your FastAPI app Dockerfile
-    ports:
-      - "8000:8000"  # Adjust port if needed
+      - "8000:8000"
     depends_on:
       - qdrant
     networks:
       - my_network
+  qdrant:
+    image: qdrant/qdrant:latest
+    restart: always
+    container_name: qdrant
+    ports:
+      - 6333:6333
+      - 6334:6334
+    expose:
+      - 6333
+      - 6334
+      - 6335
+    configs:
+      - source: qdrant_config
+        target: /qdrant/config/deployment.yaml
+    volumes:
+      - qdrant_storage:/qdrant_data
+    networks:
+      - my_network
+
+configs:
+  qdrant_config:
+    content: |
+      log_level: INFO  
+
+volumes:
+  qdrant_storage:
 
 networks:
-  my_network:
-    driver: bridge
+  my_network:
