+++ main.py
@@ -7,31 +7,31 @@ from openai import OpenAI
 from qdrant_client import QdrantClient
 from fastapi.responses import JSONResponse
 from typing import Optional
-import security
-import processing_docs
-import utils
+import security.security as security
+import lib.processing_docs as processing_docs
+import lib.utils as utils
 from docx import Document
 import io
 
-
 # Load the configuration file
-config = utils.load_config('config.yaml')
-tag = utils.load_json('tags.json') 
-tags = [tag[t] for t in tag]
-# print(tags)
-JWT_SECRET = config['secure']['SECRET_KEY']
+config = utils.load_config('./config.yaml')
 
-# Initialize Qdrant client with the host and port
+# Load documentation of the API development and show the information when is deployed
+tag = utils.load_json(r'./data/tags.json') 
+tags = [tag[t] for t in tag]
+
+# Initialize Qdrant client with the host and port according to the configuration file
 client_vdb = QdrantClient(config['qdrant']['host'], port=config['qdrant']['port'])
 
-# Initialize OpenAI client with your API key
+# Initialize OpenAI client with your API key according to the configuration file
 client_ia = OpenAI(api_key=config['openai']['key'])
 
-
+# Create a FastAPI instance with the title and tags
 app = FastAPI(title="AI Engineer Test to create", openapi_tags= tags)
 
 origins = ["*"]
 
+# Add CORS middleware to allow cross-origin requests
 app.add_middleware(
     CORSMiddleware,
     allow_origins=origins,
@@ -40,16 +40,43 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
+
+# Route for the home page
 @app.get("/", tags=["Home"])
 async def home():
-        return {'API created by Fabian Coy -2024 AI Engineer Test': 'API to load, upload data to QDRANT and get answers from OpenAI'}
+    """
+    Check if the API is running correctly.
+
+    Returns:
+        dict: A dictionary indicating that the API is running correctly.
+    """
+    return {'API created by Fabian Coy -2024 AI Engineer Test': 'API to load, upload data to QDRANT and get answers from OpenAI'}
 
 # Route for uploading a .docx file
 @app.post("/upload/", tags=["Upload_file"])
 async def upload_file(file: UploadFile = File(...), id: Optional[int] = 0, current_user: security.User = Depends(security.get_current_active_user)):
+    """
+    Upload or update a document in Qdrant vector database.
+
+    Args:
+        file (UploadFile): The .docx file to be uploaded.
+        id (Optional[int]): The ID of the document (default is 0) when the function is upload a new document. If you want to update a document, you need to provide the ID of the document.
+        current_user (security.User): The current authenticated user.
+
+    Returns:
+        JSONResponse: A JSON response indicating the status of the upload process.
+    """
+    # check if the file is a .docx file 
     if not file.filename.lower().endswith('.docx'):
         raise HTTPException(status_code=400, detail="Invalid file type. Please upload a .docx file")
+    # check if the ID is valid
+    if id != 0:    
+        information = client_vdb.get_collection(collection_name=config['qdrant']['collection'])
+        last_id = information.vectors_count
+        if id > last_id:
+            raise HTTPException(status_code=400, detail=f"Invalid ID. The last ID is {last_id}")
 
+    # read the contents of the file, create embeddings, and upload the document to Qdrant
     try:
         contents = await file.read()
         doc = Document(io.BytesIO(contents))
@@ -61,16 +88,37 @@ async def upload_file(file: UploadFile = File(...), id: Optional[int] = 0, curre
 
     except Exception:
         raise HTTPException(status_code=500, detail=f"Error processing file: {file.filename}")
-
+    # return the filename and the status of the upload process
     return JSONResponse(content={"the filename": file.filename, "process created": outcome_upload}, status_code=200)
 
+# Route to search the best document based on the prompt and create an answer using LLM model 
 @app.get("/search/{prompt}", tags=["Search"])
 async def read_words(prompt: str = Path(..., max_length=config['api']['max_length_prompt']), current_user: security.User = Depends(security.get_current_active_user)):
-    response = processing_docs.get_answer_llm(config, client_vdb, client_ia, prompt)
-    return {'response': response}
+    """
+    Search for words or phrases in the documents stored in Qdrant vector database.
 
+    Args:
+        prompt (str): The search prompt or query.
+        current_user (security.User): The current authenticated user.
+
+    Returns:
+        JSONResponse: A JSON response indicating the LLM outcome using the prompt and document.
+    """
+    response = processing_docs.get_answer_llm(config, client_vdb, client_ia, prompt)
+    return JSONResponse(content={"LLM answer": response}, status_code=200)
+
+# Route to generate a token for accessing the API securely
 @app.post("/token", response_model=security.Token, tags=["Generate Token"])
 async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
+    """
+    Generate a token for accessing the API securely.
+
+    Args:
+        form_data (OAuth2PasswordRequestForm): The form data containing the username and password.
+
+    Returns:
+        dict: A dictionary containing the access token and token type.
+    """
     user = security.authenticate_user(security.db_dummy, form_data.username, form_data.password)
     if not user:
         raise HTTPException(
